bugfix for NOBODY bug
when we use CURLOPT_NOBODY, sometimes the server responds with a 404 error, although the content should be there. The previous workaround was to add CURLOPT_HTTPGET; this only worked, however, because CURLOPT_HTTPGET sets CURLOPT_NOBODY to false (see https://curl.se/libcurl/c/CURLOPT_HTTPGET.html ), so we could also just omit the CURLOPT_NOBODY completely.
We now try CURLOPT_NOBODY, and if that fails and returns a 404, we then make a new request without CURLOPT_NOBODY.